test[#119]: response로 받는 데이터 타입이 다르기 때문에 다르게 테스트
